Remove more C bindings in favor of Rust functions.
This includes removing ykpiv_set_mgmkey, which was a very simple alias
for ykpiv_set_mgmkey2, so we can just retain the one binding.

This also includes removing the binding for ykpiv_sign_data2, which
seems to not actually be defined in upstream's C library at all, so
actually trying to call it would result in a segfault. It seems to not
be referenced anywhere either, so ykpiv_sign_data seems to be sufficient
to implement the full functionality upstream had in mind.