Add library functions for DES challenges.
We need to be able to encrypt and decrypt with DES in a way compatible
with upstream's library. So, to ensure this is the case, split out that
logic into library functions, and add tests.

Note that the tests don't currently pass, because the way we were doing
encryption / decryption was broken.