Add basic unit tests for the "change_puk" command.
This also involved making an error message more clear.

Under the hood, many of these verify/change type functions are
implemented with one common set of code. For now, this unit test
duplicates the mocks used to perform the test. But, in the future it may
make the most sense to combine them similarly to how we do in the
implementation.