Move get_version mock implementation.
The goal is to move the various mocks we need to src/tests/piv/hal.rs,
which is really an implementation file anyway. This should leave the
actual tests, in src/tests/piv/handle.rs, much more readable.

The next step is to move over a generic change handler, with support for
the various changes as well as the "retry counter" logic, so we can
properly test e.g. unblock_pin or reset or ...