Add an initial implementation of read_certificate.
This is much simpler than the implementation we used to have, and I
*think* it is correct / what the final implementation will look like.
However, it doesn't seem to work with the keys stored via the "generate"
command (neither does upstream's implementation, so this doesn't seem to
be a bug), so it can't be tested / verified yet.

The next step is to add a function like import-certificate which can be
used to store a real X.509 certificate on the device, at which point
we should be able to verify + test this function.