Add encrypt / decrypt functionality, and a "test_decrypt" command.
This is analogous to upstream's "test-decipher" command, but it is more
generally useful (e.g., we can use this encryption to wrap keys for
storing encrypted data and decrypting it with the hardware device).

Unfortunately, this requires a patch to the Rust openssl bindings,
because it depends on a function which previously was not included in
the bindings.