Derive deserialize and serialize for various types.
For many of these types, it's clearly useful to be able to put them in
e.g. configuration data persisted to disk, so allowing them to be
serialized easily is a plus.

Since serde is a dependency anyway for PIV recording functionality, this
doesn't add extra overhead to the library.