Reconfigure the crate to use features to enable OTP and PIV code.
This lets users use just one,  without pulling in all the dependencies
of the other one.

This is a little bit incomplete. The `piv-tool` binary sort of "forces"
the PIV feature to be enabled on a "cargo build". To fix this, a future
commit will split the binary out into a different crate, and make this
repository a workspace with the library and tool crates.