First attempt at fixing a threading bug on connect/disconnect.
If we use handles to connect / disconnect from a single device in
multiple threads, one thread disconnecting the device causes the other
threads to get a "card is being reset" error.

This commit attempts to fix this by keeping track of how many handles we
have per device, and only resetting the card when the final one is
destructed.

Unfortunately, this doesn't seem to 100% solve the issue.